Combine AutoSize logic into Container.
